requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
sqlalchemy_utils
werkzeug
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
